refactor: unify naming conventions and reorganize files
