+++ app/src/main/java/com/example/gitgrow/ui/theme/AnimationConfig.kt
@@ -5,6 +5,4 @@ package com.example.gitgrow.ui.theme
 data object AnimationConfig {
     const val SlideInDurationMillis = 700
     const val SlideOutDurationMillis = 700
-    const val FadeInDurationMillis = 300
-    const val FadeOutDurationMillis = 300
 }
